requirements.txt
asgiref==3.11.0
Django==6.0
django-browser-reload==1.21.0
django-tailwind==4.4.2
honcho==2.0.0
pillow==12.0.0
pytailwindcss==0.3.0
python-barcode==0.16.1
sqlparse==0.5.5
